feat: makeCalEvents
