+++ src/common/index.ts
@@ -1,3 +1,4 @@
 export type { default as CalendarEvent } from './CalendarEvent'
 export { default as Swal } from './AppSwal'
 export { default as unimplemented } from './unimplemented'
+export { default as makeCalEvents } from './makeCalEvents'
