+++ src/components/ShowAppointments.vue
@@ -2,11 +2,9 @@
 import { computed } from 'vue'
 import dayjs, { type Dayjs } from 'dayjs'
 import { useI18n } from 'vue-i18n'
-import groupBy from 'lodash.groupby'
-import { Swal } from '@/common'
-import useAppointments from '@/stores/appointments'
 import type { Place, Appointment, PendingAppointment } from '@/models'
-import type { CalendarEvent } from '@/common'
+import { Swal, makeCalEvents } from '@/common'
+import useAppointments from '@/stores/appointments'
 import Errors from '@/errors'
 
 const props = defineProps<{
@@ -18,45 +16,13 @@ const props = defineProps<{
 const { t } = useI18n()
 
 const appointments = useAppointments()
-
-function u(text: string) {
-  return `<u>${text}</u>`
-}
-
-function mergeAppointmentsWithPendings(
-  arg: Array<Appointment & PendingAppointment>
-) {
-  return arg.reduce((acc: string[], curr) => {
-    const newElements = (curr.owner ? [u(curr.owner)] : curr.users) as string[]
-
-    return [...acc, ...newElements]
-  }, [])
-}
-
-function isAppPending(a: Appointment | PendingAppointment): boolean {
-  return !!(a as PendingAppointment).owner
-}
-
-const events = computed<CalendarEvent[]>(() => {
-  const grouped = groupBy(
-    [...props.appointments, ...props.pendingAppointments],
-    e => e.place.id + e.datetime
-  )
-
-  return Object.values(grouped).map(appointments => {
-    const [appointment] = appointments
-    const isPending = appointments.some(a => isAppPending(a))
-    const users = mergeAppointmentsWithPendings(appointments)
-
-    return {
-      id: appointment.id,
-      content: users,
-      datetime: dayjs(appointment.datetime),
-      title: !props.place ? appointment.place.name : '',
-      pending: isPending,
-    } as CalendarEvent
+const events = computed(() =>
+  makeCalEvents({
+    appointments: props.appointments,
+    pendingAppointments: props.pendingAppointments,
+    place: props.place,
   })
-})
+)
 
 async function missingPlaceSwal() {
   const dontShowAgainKey = 'dont-show-missing-place-again'
